Seperate upload class and API
there are assumptions around how sub-imports are used in shared context
by separating this out, we can more cleaning use Upload class in shared context without affecting flask route
